+++ R/vcovPC.R
@@ -21,6 +21,9 @@
 vcovPC.prais <- function(x, pairwise = FALSE, ...) {
 
   coeffs <- x$coefficients
+  # Coefficients of linearly dependent variables are NA and are omitted, as in
+  # the covariance matrix of an object of class 'lm'
+  coeffs <- coeffs[!is.na(coeffs)]
   if (length(coeffs) > 0) {
     x_names <- names(coeffs)
     rho <- x$rho[NROW(x$rho), ]